Preliminary edits from NTuplers into TelescopingJets class
