+++ Ana_EventGeneration/NTupler.cc
@@ -620,537 +620,7 @@ std::vector<fastjet::PseudoJet> ToyCalorimeter(std::vector<fastjet::PseudoJet> t
     
     return cellParticles;
 }
-    
-///=========================================
-/// Telescoping Pruning
-///=========================================
-double T_Pruning(fastjet::PseudoJet& input, double minDCut, double maxDCut, int numDCuts) {
-    double zCut = 0.1; // Single choice of zcut but can be further telescoped.
-    std::vector<double> telescopingMasses;
 
-    double dCutStep = (maxDCut - minDCut) / (numDCuts);
-
-    //  Obtain a set of jet masses by stepping through the DCut
-    //  pruning parameter.
-    for (double dCut = minDCut; dCut < maxDCut + dCutStep; dCut += dCutStep) {
-        fastjet::Pruner pruner(fastjet::cambridge_algorithm, zCut, dCut);
-        fastjet::PseudoJet prunedJet = pruner(input);
-        double mass = prunedJet.m();
-        if (mass > M0) telescopingMasses.push_back(mass);
-    }
-
-    
-    if (!telescopingMasses.empty()) return getVolatility(telescopingMasses);
-    
-    std::cout << "WARNING zero entries for T_Pruning!   minDCut: " << minDCut <<
-	"   maxDCut: " << maxDCut <<
-	"   numDCuts: " << numDCuts <<
-	"   input mass: " << input.m() << std::endl;
-    return -1;
-}
-
-
-///=========================================
-/// Telescoping Trimming
-///=========================================
-double T_Trimming(fastjet::PseudoJet& input, double minFCut, double maxFCut, int numFCuts) {
-    //  The subjet radius used should depend on the input jet's
-    //  pT--higher pT should have a smaller subjet radius.
-    double Rfilt = 0.1; 
-    if (input.pt() > 500) Rfilt = 0.05;
-    
-    std::vector<double> telescopingMasses;
-    
-    double fCutStep = (maxFCut - minFCut) / (numFCuts);
-
-    //  Obtain a set of jet masses by stepping through the f trimming parameter.
-    for (double fCut = minFCut; fCut < maxFCut + fCutStep; fCut += fCutStep) {
-        fastjet::Filter trimmer(Rfilt, fastjet::SelectorPtFractionMin(fCut));
-        fastjet::PseudoJet trimmedJet = trimmer(input);
-        double mass = trimmedJet.m();
-        if (mass > M0) telescopingMasses.push_back(mass);
-    }
-
-    if (!telescopingMasses.empty()) return getVolatility(telescopingMasses);
-    
-    std::cout << "WARNING zero entries for T_Trimming!   minFCut: "<< minFCut <<
-	"   maxFCut: "<< maxFCut <<
-	"   numFCuts: "<< numFCuts <<
-	"   input mass: " << input.m() << std::endl;
-    return -1;
-}
-
-///=========================================
-/// Telescoping reclustering
-/// for kt, set algorithm = 0
-/// for antiki, set algorithm = 2
-///=========================================
-double T_Reclustering(fastjet::PseudoJet& input, int algorithm, double minRadius, double maxRadius, int numRadii) {
-    std::vector<double> telescopingMass; 
-    double deltaR = (maxRadius - minRadius) / (numRadii);
-
-    for (double r = minRadius; r < maxRadius + deltaR; r += deltaR) {
-        fastjet::JetDefinition TjetDefinition(fastjet::JetAlgorithm(algorithm), r);
-        fastjet::ClusterSequence TClusterSequence(input.constituents(), TjetDefinition);
-        std::vector<fastjet::PseudoJet> recoTJets = sorted_by_pt(TClusterSequence.inclusive_jets());
-	
-        if (recoTJets.empty()) {
-            std::cout << "WARNING! Recluster number of subjets is " << recoTJets.size() << std::endl;
-            continue;
-        }
-        if (recoTJets.size() == 1) {
-            Tsubjet1.SetPxPyPzE(recoTJets[0].px(), recoTJets[0].py(), recoTJets[0].pz(), recoTJets[0].e());
-            double mass = Tsubjet1.M();
-            if (mass > M0) telescopingMass.push_back(mass);
-        }
-        else if (recoTJets.size() >= 2) {
-            Tsubjet1.SetPxPyPzE(recoTJets[0].px(), recoTJets[0].py(), recoTJets[0].pz(), recoTJets[0].e());
-            Tsubjet2.SetPxPyPzE(recoTJets[1].px(), recoTJets[1].py(), recoTJets[1].pz(), recoTJets[1].e());
-            double mass = (Tsubjet1 + Tsubjet2).M();
-            if (mass > M0) telescopingMass.push_back(mass);
-        }
-    }
-    
-    if (!telescopingMass.empty()) return getVolatility(telescopingMass);
-    
-    std::cout << "WARNING: zero entries for T_reclustering!   Algorithm" << algorithm <<
-	"   minRadius: " << minRadius <<
-	"   maxRadius: " << maxRadius <<
-	"   numRadii: " << numRadii <<  std::endl;
-    return -1;
-}
-
-///=========================================
-/// Telescoping Subjet
-///=========================================
-TSub TNSubjet(fastjet::PseudoJet& input, unsigned int numSubjets, double minRadius, double maxRadius, int numRadii, int stepScale) {
-    TSub result;
-    
-    double beta = 1.0;    
-    fastjet::contrib::UnnormalizedMeasure nSubMeasure(beta);
-    //fastjet::contrib::Nsubjettiness nSubjettiness(numSubjets, fastjet::contrib::OnePass_KT_Axes(), nSubMeasure);
-    fastjet::contrib::Nsubjettiness nSubjettiness(numSubjets, fastjet::contrib::WTA_KT_Axes(), nSubMeasure);
-
-    double tauN = nSubjettiness.result(input);
-    std::vector<fastjet::PseudoJet> tauAxes = nSubjettiness.currentAxes();
-    std::vector<TLorentzVector> pTauAxes(numSubjets);
-
-    for (unsigned int i = 0; i < numSubjets; ++i) {
-	pTauAxes[i].SetPxPyPzE(tauAxes[i].px(), tauAxes[i].py(), tauAxes[i].pz(), tauAxes[i].e());
-    }
-
-    std::vector<double> anglesBetweenTauAxes;
-    for (unsigned int i = 0; i < numSubjets; ++i) {
-	for (unsigned int j = i + 1; j < numSubjets; ++j) {
-	    anglesBetweenTauAxes.push_back(pTauAxes[i].DeltaR(pTauAxes[j]));
-	}
-    }
-
-    if (!anglesBetweenTauAxes.empty()) result.minAngle = *(std::min_element(anglesBetweenTauAxes.cbegin(), anglesBetweenTauAxes.cend()));
-    
-    std::vector<fastjet::PseudoJet> constituents = input.constituents();
-    std::vector<std::vector<std::pair<TLorentzVector, double>>> sortedConstituents(numSubjets);
-    
-    for (auto const &constituent: constituents) {
-	TLorentzVector pConstituent(constituent.px(), constituent.py(), constituent.pz(), constituent.e());
-
-	std::vector<double> distanceToTauAxes(numSubjets);
-	for (unsigned int i = 0; i < numSubjets; ++i) {
-	    distanceToTauAxes[i] = pConstituent.DeltaR(pTauAxes[i]);
-	}
-
-	auto minDistanceToTauAxesIt = std::min_element(distanceToTauAxes.cbegin(), distanceToTauAxes.cend());
-	double minDistanceToTauAxes = *minDistanceToTauAxesIt;
-	unsigned int minDistanceToTauAxesIndex = std::distance(distanceToTauAxes.cbegin(), minDistanceToTauAxesIt);
-
-	sortedConstituents[minDistanceToTauAxesIndex].push_back({pConstituent, minDistanceToTauAxes});
-    }
-    
-    for (auto &subjetConstituents : sortedConstituents) {
-	std::sort(subjetConstituents.begin(), subjetConstituents.end(), [](const std::pair<TLorentzVector, double> &pair1,
-									   const std::pair<TLorentzVector, double> &pair2) {
-		      return pair1.second < pair2.second;
-		  });
-    }
-    
-    std::vector<TLorentzVector> TSubjets(numSubjets);
-    std::vector<double> telescopingMasses;
-    std::vector<double> telescopingPT;
-
-    double deltaR = -999;
-    if (stepScale == 0) deltaR = (maxRadius - minRadius) / (numRadii);
-    else if (stepScale == 1) deltaR = log10(maxRadius / minRadius) / (numRadii);
-
-    for (int i = 0; i <= numRadii; ++i) {
-	double r = -999;
-	if (stepScale == 0) r = minRadius + i * deltaR;
-	else if (stepScale == 1) r = minRadius * pow(10, i * deltaR);
-
-	for (unsigned int j = 0; j < sortedConstituents.size(); ++j) {
-	    for (auto it = sortedConstituents[j].begin(); it != sortedConstituents[j].end(); ++it) {
-		if (it->second <= r) {
-		    TSubjets[j] += it->first;
-		    sortedConstituents[j].erase(it);
-		    --it;
-		}
-		else break;
-	    }
-	}
-	
-	TLorentzVector sumTSubjet;
-	for (auto const &TSubjet : TSubjets) {
-	    sumTSubjet += TSubjet;
-	}
-	
-	if (sumTSubjet.M() > M0) {
-	    telescopingMasses.push_back(sumTSubjet.M());
-	    telescopingPT.push_back(sumTSubjet.Pt());
-	}
-    }
-
-    if (!telescopingMasses.empty()) {
-	result.massVolatility = getVolatility(telescopingMasses);
-	result.pTVolatility = getVolatility(telescopingPT);
-	result.masses = telescopingMasses;
-    }
-    else {
-	std::cout << "WARNING zero entries for TNSubjet!   numSubjets: " << numSubjets <<
-	    "   minRadius: " << minRadius <<
-	    "   maxRadius: " << maxRadius <<
-	    "   numRadii: " << numRadii <<
-	    "   input mass: " << input.m() << std::endl;
-    }
-    return result;
-}
-
-
-T2Sub T2Subjet(fastjet::PseudoJet& input, double minRadius, double maxRadius, int numRadii, int stepScale) {
-    T2Sub result;
-    
-    double beta = 1.0;
-    fastjet::contrib::UnnormalizedMeasure nSubMeasure(beta);
-    //fastjet::contrib::Nsubjettiness nSubjettiness(2, fastjet::contrib::OnePass_KT_Axes(), nSubMeasure);
-    fastjet::contrib::Nsubjettiness nSubjettiness(2, fastjet::contrib::WTA_KT_Axes(), nSubMeasure);
-
-    double tau2 = nSubjettiness.result(input);
-    std::vector<fastjet::PseudoJet> tau2axes = nSubjettiness.currentAxes();
-
-    std::vector<TLorentzVector> pTauAxes(2);
-    for (unsigned int i = 0; i < 2; ++i) {
-	pTauAxes[i].SetPxPyPzE(tau2axes[i].px(), tau2axes[i].py(), tau2axes[i].pz(), tau2axes[i].e());
-    }
-
-    result.minAngle = pTauAxes[0].DeltaR(pTauAxes[1]);
-
-    std::vector<fastjet::PseudoJet> constituents = input.constituents();
-    std::vector<std::vector<std::pair<TLorentzVector, double>>> sortedConstituents(2);
-
-    for (auto const &constituent : constituents) {
-	TLorentzVector pConstituent(constituent.px(), constituent.py(), constituent.pz(), constituent.e());
-
-	std::vector<double> distanceToTauAxes(2);
-	for (unsigned int i = 0; i < 2; ++i) {
-	    distanceToTauAxes[i] = pConstituent.DeltaR(pTauAxes[i]);
-	}
-
-	auto minDistanceToTauAxesIt = std::min_element(distanceToTauAxes.cbegin(), distanceToTauAxes.cend());
-	double minDistanceToTauAxes = *minDistanceToTauAxesIt;
-	unsigned int minDistanceToTauAxesIndex = std::distance(distanceToTauAxes.cbegin(), minDistanceToTauAxesIt);
-
-	sortedConstituents[minDistanceToTauAxesIndex].push_back({pConstituent, minDistanceToTauAxes});
-    }
-    
-    for (auto &subjetConstituents : sortedConstituents) {
-	std::sort(subjetConstituents.begin(), subjetConstituents.end(), [](const std::pair<TLorentzVector, double> &pair1,
-									   const std::pair<TLorentzVector, double> &pair2) {
-		      return pair1.second < pair2.second;
-		  });
-    }
-    
-    std::vector<TLorentzVector> TSubjets(2);
-    std::vector<std::vector<double>> telescopingMasses(3);
-    std::vector<std::vector<double>> telescopingPTs(3);
-    std::vector<double> masses(3);
-    std::vector<double> pTs(3);
-
-    double deltaR = -999;
-    if (stepScale == 0) deltaR = (maxRadius - minRadius) / (numRadii);
-    else if (stepScale == 1) deltaR = log10(maxRadius / minRadius) / (numRadii);
-
-    for (int i = 0; i <= numRadii; ++i) {
-	double r = -999;
-	if (stepScale == 0) r = minRadius + i * deltaR;
-	else if (stepScale == 1) r = minRadius * pow(10, i * deltaR);
-
-	for (unsigned int j = 0; j < sortedConstituents.size(); ++j) {
-	    for (auto it = sortedConstituents[j].begin(); it != sortedConstituents[j].end(); ++it) {
-		if (it->second <= r) {
-		    TSubjets[j] += it->first;
-		    sortedConstituents[j].erase(it);
-		    --it;
-		}
-		else break;
-	    }
-	}
-	masses[0] = (TSubjets[0] + TSubjets[1]).M();
-	masses[1] = TSubjets[0].M();
-	masses[2] = TSubjets[1].M();
-
-	pTs[0] = (TSubjets[0] + TSubjets[1]).Pt();
-	pTs[1] = TSubjets[0].Pt();
-	pTs[2] = TSubjets[1].Pt();
-
-	for (unsigned int i = 0; i < masses.size(); ++i) {
-	    telescopingMasses[i].push_back(masses[i]);
-	    telescopingPTs[i].push_back(pTs[i]);
-	}
-    }
-    
-    if (!telescopingMasses[0].empty() && !telescopingMasses[1].empty() && !telescopingMasses[2].empty()) {
-	result.massVolatility = getVolatility(telescopingMasses[0]);
-	result.pTVolatility = getVolatility(telescopingPTs[0]);
-	result.masses = telescopingMasses[0];
-
-	std::vector<double> residualWMass = masses;
-
-	std::transform(residualWMass.begin() + 1, residualWMass.end(), residualWMass.begin() + 1,
-		       bind2nd(std::minus<double>(), massW));
-	std::transform(residualWMass.begin() + 1, residualWMass.end(), residualWMass.begin() + 1,
-		       static_cast<double (*)(double)>(&std::abs));
-
-	auto wMassPredictionIt = std::min_element(residualWMass.cbegin() + 1, residualWMass.cend());
-	unsigned int wMassPredictionIndex = std::distance(residualWMass.cbegin(), wMassPredictionIt);
-
-	result.wMass = masses[wMassPredictionIndex];
-	result.wMassVolatility = getVolatility(telescopingMasses[wMassPredictionIndex]);
-    }
-
-    else {
-	std::cout << "WARNING zero entries for T2Subjet!   minRadius: " << minRadius <<
-	    "   maxRadius: " << maxRadius <<
-	    "   numRadii: " << numRadii <<
-	    "   input mass: " << input.m() << std::endl;
-    }
-    return result;
-}
-
-T3Sub T3Subjet(fastjet::PseudoJet& input, double minRadius, double maxRadius, int numRadii, int stepScale) {
-    T3Sub result;
-
-    double beta = 1.0;
-    fastjet::contrib::UnnormalizedMeasure nSubMeasure(beta);
-    //fastjet::contrib::Nsubjettiness nSubjettiness(3, fastjet::contrib::OnePass_KT_Axes(), nSubMeasure);
-    fastjet::contrib::Nsubjettiness nSubjettiness(3, fastjet::contrib::WTA_KT_Axes(), nSubMeasure);
-    double tau3 = nSubjettiness.result(input);
-    std::vector<fastjet::PseudoJet> tau3axes = nSubjettiness.currentAxes();
-
-    std::vector<TLorentzVector> pTauAxes(3);
-    for (unsigned int i = 0; i < 3; ++i) {
-	pTauAxes[i].SetPxPyPzE(tau3axes[i].px(), tau3axes[i].py(), tau3axes[i].pz(), tau3axes[i].e());
-    }
-    
-    std::vector<double> anglesBetweenTauAxes;
-    for (unsigned int i = 0; i < 3; ++i) {
-	for (unsigned int j = i + 1; j < 3; ++j) {
-	    anglesBetweenTauAxes.push_back(pTauAxes[i].DeltaR(pTauAxes[j]));
-	}
-    }
-
-    std::vector<double> tempAnglesBetweenTauAxes = anglesBetweenTauAxes;
-    std::sort(tempAnglesBetweenTauAxes.begin(), tempAnglesBetweenTauAxes.end());
-
-    result.minAngle = tempAnglesBetweenTauAxes[0];
-    result.midAngle = tempAnglesBetweenTauAxes[1];
-    result.maxAngle = tempAnglesBetweenTauAxes[2];
-
-    std::vector<fastjet::PseudoJet> constituents = input.constituents();
-    std::vector<std::vector<std::pair<TLorentzVector, double>>> sortedConstituents(3);
-
-    for (auto const &constituent: constituents) {
-        TLorentzVector pConstituent(constituent.px(), constituent.py(), constituent.pz(), constituent.e());
-
-        std::vector<double> distanceToTauAxes(3);
-        for (unsigned int i = 0; i < 3; ++i) {
-            distanceToTauAxes[i] = pConstituent.DeltaR(pTauAxes[i]);
-        }
-
-	auto minDistanceToTauAxesIt = std::min_element(distanceToTauAxes.cbegin(), distanceToTauAxes.cend());
-        double minDistanceToTauAxes = *minDistanceToTauAxesIt;
-        unsigned int minDistanceToTauAxesIndex = std::distance(distanceToTauAxes.cbegin(), minDistanceToTauAxesIt);
-
-        sortedConstituents[minDistanceToTauAxesIndex].push_back({pConstituent, minDistanceToTauAxes});
-    }
-
-    for (auto &subjetConstituents : sortedConstituents) {
-	std::sort(subjetConstituents.begin(), subjetConstituents.end(), [](const std::pair<TLorentzVector, double> &pair1, const std::pair<TLorentzVector, double> &pair2) {
-                return pair1.second < pair2.second;
-            });
-    }
-
-    std::vector<TLorentzVector> TSubjets(3);
-    std::vector<std::vector<double>> telescopingMasses(4);
-    std::vector<std::vector<double>> telescopingPTs(4);
-    std::vector<double> masses(4);
-    std::vector<double> pTs(4);
-
-    double deltaR = -999;
-    if (stepScale == 0) deltaR = (maxRadius - minRadius) / (numRadii);
-    else if (stepScale == 1) deltaR = log10(maxRadius / minRadius) / (numRadii);
-    
-    for (int i = 0; i <= numRadii; ++i) {
-	double r = -999;
-	if (stepScale == 0) r = minRadius + i * deltaR;
-	else if (stepScale == 1) r = minRadius * pow(10, i * deltaR);
-
-	for (unsigned int j = 0; j < sortedConstituents.size(); ++j) {
-	    for (auto it = sortedConstituents[j].begin(); it != sortedConstituents[j].end(); ++it) {
-		if (it->second <= r) {
-		    TSubjets[j] += it->first;
-		    sortedConstituents[j].erase(it);
-		    --it;
-		}
-		else break;
-	    }
-	}
-	
-	masses[0] = (TSubjets[0] + TSubjets[1] + TSubjets[2]).M();
-	masses[1] = (TSubjets[0] + TSubjets[1]).M();
-	masses[2] = (TSubjets[0] + TSubjets[2]).M();
-	masses[3] = (TSubjets[1] + TSubjets[2]).M();
-
-	pTs[0] = (TSubjets[0] + TSubjets[1] + TSubjets[2]).Pt();
-	pTs[1] = (TSubjets[0] + TSubjets[1]).Pt();
-	pTs[2] = (TSubjets[0] + TSubjets[2]).Pt();
-	pTs[3] = (TSubjets[1] + TSubjets[2]).Pt();
-
-
-	for (unsigned int i = 0; i < masses.size(); ++i) {
-	    telescopingMasses[i].push_back(masses[i]);
-	    telescopingPTs[i].push_back(pTs[i]);
-	}
-    }
-    
-    if (!telescopingMasses[0].empty() && !telescopingMasses[1].empty() && !telescopingMasses[2].empty() && !telescopingMasses[3].empty()) {
-        result.massVolatility = getVolatility(telescopingMasses[0]);
-	result.pTVolatility = getVolatility(telescopingPTs[0]);
-	result.masses = telescopingMasses[0];
-
-	//  TODO
-	//  Is there a better way to recover W masses?
-	std::vector<double> residualWMass = masses;
-	
-	std::transform(residualWMass.begin() + 1, residualWMass.end(), residualWMass.begin() + 1,
-		  bind2nd(std::minus<double>(), massW));
-	std::transform(residualWMass.begin() + 1, residualWMass.end(), residualWMass.begin() + 1,
-		       static_cast<double (*)(double)>(&std::abs));
-
-	auto wMassPredictionIt = std::min_element(residualWMass.cbegin() + 1, residualWMass.cend());
-	unsigned int wMassPredictionIndex = std::distance(residualWMass.cbegin(), wMassPredictionIt);
-
-	result.wMass = masses[wMassPredictionIndex];
-	result.wMassVolatility = getVolatility(telescopingMasses[wMassPredictionIndex]);
-	result.wPTVolatility = getVolatility(telescopingPTs[wMassPredictionIndex]);
-    }
-
-    else {
-        std::cout << "WARNING zero entries for T3Subjet!   minRadius: " << minRadius <<
-	    "   maxRadius: " << maxRadius <<
-	    "   numRadii: " << numRadii <<
-	    "   input mass: " << input.m() << std::endl;
-    }    
-    return result;
-}
-
-///=========================================
-/// Telescoping N-subjettiness
-///=========================================
-double T_Nsubjettiness(int N, fastjet::PseudoJet& input, double minBeta, double maxBeta, int numBetas) {
-  std::vector<double> taus;
-  double betaStep = (maxBeta - minBeta) / (numBetas);
-  
-  for (double beta = minBeta; beta < maxBeta; beta += betaStep) {
-      fastjet::contrib::UnnormalizedMeasure nSubMeasure(beta);
-      fastjet::contrib::Nsubjettiness nSub(N, fastjet::contrib::OnePass_KT_Axes(), nSubMeasure);
-      //fastjet::contrib::Nsubjettiness nsub(N, fastjet::contrib::WTA_KT_Axes(), nsubMeasure);
-      taus.push_back(nSub(input));
-  }
-  if (!taus.empty()) return getVolatility(taus);
-  
-  std::cout <<"WARNING zero entries for T_Nsubjettiness! minBeta: " << minBeta << "\tmaxBeta: "<< maxBeta << "\tnumBetas: " << numBetas << std::endl;
-  return -1;
-}
-
-double T_NsubjettinessRatio(int nNumerator, int nDemoninator, fastjet::PseudoJet& input, double minBeta, double maxBeta, int numBetas) {
-  std::vector<double> taus;
-  double betaStep = (maxBeta - minBeta) / (numBetas);
-  
-  for (double beta = minBeta; beta < maxBeta + betaStep; beta += betaStep) {
-      fastjet::contrib::UnnormalizedMeasure nsubMeasure(beta);
-      
-      fastjet::contrib::Nsubjettiness nSubNumerator(nNumerator, fastjet::contrib::WTA_KT_Axes(), nsubMeasure);
-      fastjet::contrib::Nsubjettiness nSubDenominator(nDemoninator, fastjet::contrib::WTA_KT_Axes(), nsubMeasure);
-      
-      double numerator = nSubNumerator(input);
-      double denominator = nSubDenominator(input);
-      
-      if (denominator != 0) taus.push_back(numerator / denominator);
-      else taus.push_back(-1.0);
-  }
-  
-  if (!taus.empty()) return getVolatility(taus);
-  
-  std::cout <<"WARNING: zero entries for T_NsubjettinessRatio "<< minBeta <<" "<< maxBeta <<" "<< numBetas <<" "<<  std::endl;
-  return -1;
-}
-
-
-///=========================================
-/// Telescoping Energy Correlators
-///=========================================
-double T_EnergyCorrelator_C2(fastjet::PseudoJet& input, double minBeta, double maxBeta, int numBetas) {
-    std::vector<double> ecfs; ecfs.clear();
-    for (int i = 0; i < numBetas; i++) {
-	double beta = minBeta + i*(maxBeta - minBeta)/(numBetas-1);
-	fastjet::contrib::EnergyCorrelatorC2 ecf(beta);
-	ecfs.push_back(ecf(input));
-    }
-    // getVolatility function provided by TelescopingJets
-    if(ecfs.size()>0)
-        return getVolatility(ecfs);
-    
-    std::cout <<"WARNING: zero entries for T_EnergyCorrelator_C2 "<< minBeta <<" "<< maxBeta <<" "<< numBetas <<" "<<  std::endl;
-    return -1;   
-}
-
-double T_EnergyCorrelator_D2(fastjet::PseudoJet& input, double minBeta, double maxBeta, int numBetas) {
-    std::vector<double> ecfs; ecfs.clear();
-    for (int i = 0; i < numBetas; i++) {
-	double beta = minBeta + i*(maxBeta - minBeta)/(numBetas-1);
-	fastjet::contrib::EnergyCorrelatorD2 ecf(beta);
-	ecfs.push_back(ecf(input));
-    }
-    // getVolatility function provided by TelescopingJets
-    if(ecfs.size()>0)
-        return getVolatility(ecfs);
-    
-    std::cout <<"WARNING: zero entries for T_EnergyCorrelator_C2 "<< minBeta <<" "<< maxBeta <<" "<< numBetas <<" "<<  std::endl;
-    return -1;   
-}
-
-double T_EnergyCorrelator_C3(fastjet::PseudoJet& input, double minBeta, double maxBeta, int numBetas) {
-    std::vector<double> ecfs; ecfs.clear();
-    for (int i = 0; i < numBetas; i++) {
-	double beta = minBeta + i*(maxBeta - minBeta)/(numBetas-1);
-	fastjet::contrib::EnergyCorrelatorDoubleRatio ecf(3, beta);
-	ecfs.push_back(ecf(input));
-    }
-    // getVolatility function provided by TelescopingJets
-    if(ecfs.size()>0)
-        return getVolatility(ecfs);
-    
-    std::cout <<"WARNING: zero entries for T_EnergyCorrelator_C3 "<< minBeta <<" "<< maxBeta <<" "<< numBetas <<" "<<  std::endl;
-    return -1;
-}
 
 ///========================================
 int GetJetTruthFlavor(TLorentzVector tempJet,
